:recycle: Make serializeQueryString a global helper
